Server/requirements.txt
fastapi 
uvicorn 
pillow 
opencv-python 
torch 
torchvision 
segmentation-models-pytorch 
ultralytics 
python-multipart